# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Money API",
        "description": "API pour les services financiers mobile",
        "contact": {
            "email": "support@orangemoney.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ompay-4mgy.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/v1/wallet/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Obtenir le solde du compte",
                "operationId": "dd3af443c2049b4ee08a8bbb6b009c7b",
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/wallet/deposit": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Dépôt d'argent",
                "operationId": "bcff6bc99dec564c873ec30fee98ec59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}